Refactor heat.py
