+++ xija/component/heat/earth_vis_grid_nside32.fits.gz
